tro 最新代码合并! flyway sql 优化; 探针状态模式增加; 指标数据存储数据结构修改; 性能分析数据定时清理; 探针命令过期时间; 其他优化等...

+++ takin-web-common/src/main/java/com/pamirs/takin/common/util/Snowflake.java
@@ -68,6 +68,6 @@ public class Snowflake {
             referenceTime = currentTime;
         }
 
-        return currentTime << NODE_SHIFT << SEQ_SHIFT | node << SEQ_SHIFT | counter;
+        return currentTime << NODE_SHIFT << SEQ_SHIFT | (long)node << SEQ_SHIFT | counter;
     }
 }
